Describe dataset

codebook() displays a compact overview of the data frame. It will have important characteristics of the data frame including type, label, observation number, and missing values as well as percentage of missing values.

codebook(data)

Arguments

data
A data.frame

Examples

Here we show the usage using the dataset iris.

codebook(iris)
#>         Codebook
#>    Dataset's Name : `iris`
#>   Dataset's Label : 
#>        Vars : 5
#>         Obs : 150
#>  + -- ------------ + ----- ------- --- -- ----- +
#>  | No     Variable | Label    Type Obs NA NA(%) |
#>  + -- ------------ + ----- ------- --- -- ----- +
#>  |  1 Sepal.Length |       numeric 150  0   0.0 |
#>  |  2  Sepal.Width |       numeric 150  0   0.0 |
#>  |  3 Petal.Length |       numeric 150  0   0.0 |
#>  |  4  Petal.Width |       numeric 150  0   0.0 |
#>  |  5      Species |        factor 150  0   0.0 |
#>  + -- ------------ + ----- ------- --- -- ----- +